+++ srcs/headers/shapes.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   shapes.h                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: craimond <bomboclat@bidol.juis>            +#+  +:+       +#+        */
+/*   By: egualand <egualand@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/23 21:47:12 by craimond          #+#    #+#             */
-/*   Updated: 2024/04/06 16:19:52 by craimond         ###   ########.fr       */
+/*   Updated: 2024/04/10 17:53:13 by egualand         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@
 typedef struct s_sphere
 {
 	t_point		center;
-	float		radius;
+	double		radius;
 }	t_sphere;
 
 typedef struct s_plane
@@ -32,9 +32,9 @@ typedef struct s_cylinder
 {
 	t_point		center;
 	t_vector	direction;
-	float		radius;
-	float		sqr_radius;
-	float		half_height;
+	double		radius;
+	double		sqr_radius;
+	double		half_height;
 	t_point		top_cap_center;
 	t_point		bottom_cap_center;
 }	t_cylinder;
